library/akdof_shared/pyproject.toml
[project]
name = "akdof_shared"
version = "0.1.0"
description = "An internal akdof-monorepo package for code that is shared across multiple projects."
authors = [
    { name = "Connor Edick", email = "connor.edick@alaska.gov" }
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.10",
    "aiodns>=3.0.0",
    "geopandas>=1.0.1",
    "geomet>=1.0.0",
    "keyring>=24.3.0",
    "keyrings-cryptfile>=1.3.9",
    "pandas>=2.2.3",
    "pydantic>=2.10.3",
    "requests>=2.32.3"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["akdof_shared"]